refactor: move queue-msg into a library
