+++ dictionary.txt
@@ -198,6 +198,7 @@ blcok
 blocksync
 blst
 boneh
+bonlulu
 bools
 builtins
 bumpp
