+++ lib/pg-queue/.sqlx/query-03b14d431114c0ef4b046b0bbf1cf0cf98df4df3d8301fa2e08454daabf273df.json
@@ -1,6 +1,6 @@
 {
   "db_name": "PostgreSQL",
-  "query": "INSERT INTO queue (item) VALUES ($1) RETURNING id",
+  "query": "INSERT INTO queue (topic, item) VALUES ($1, $2) RETURNING id",
   "describe": {
     "columns": [
       {
@@ -11,6 +11,7 @@
     ],
     "parameters": {
       "Left": [
+        "Varchar",
         "Jsonb"
       ]
     },
@@ -18,5 +19,5 @@
       false
     ]
   },
-  "hash": "58f56ca9b0ebea859b95d5f45650b00f57ac9bd459b532d0728703f51b5f743f"
+  "hash": "03b14d431114c0ef4b046b0bbf1cf0cf98df4df3d8301fa2e08454daabf273df"
 }
