+++ lib/pg-queue/.sqlx/query-7f6a46ffc2a99a2ca8c79888bb476be225e19cbfe0f61287b791ec73c634e4a8.json
@@ -1,0 +1,28 @@
+{
+  "db_name": "PostgreSQL",
+  "query": "\n            UPDATE queue\n            SET status = 'done'::status\n            WHERE id = (\n              SELECT id\n              FROM queue\n              WHERE status = 'ready'::status\n              AND topic = $1\n              ORDER BY id ASC\n              FOR UPDATE SKIP LOCKED\n              LIMIT 1\n            )\n            RETURNING id, item as \"item: Json<T>\"",
+  "describe": {
+    "columns": [
+      {
+        "ordinal": 0,
+        "name": "id",
+        "type_info": "Int8"
+      },
+      {
+        "ordinal": 1,
+        "name": "item: Json<T>",
+        "type_info": "Jsonb"
+      }
+    ],
+    "parameters": {
+      "Left": [
+        "Text"
+      ]
+    },
+    "nullable": [
+      false,
+      false
+    ]
+  },
+  "hash": "7f6a46ffc2a99a2ca8c79888bb476be225e19cbfe0f61287b791ec73c634e4a8"
+}
