+++ lib/queue-msg/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+edition = "2021"
+name    = "queue-msg"
+version = "0.1.0"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+arbitrary                = { workspace = true, optional = true, features = ["derive"] }
+frame-support-procedural = "18.0.0"
+frunk                    = "0.4.2"
+serde                    = { version = "1.0.173", features = ["derive"] }
+serde_json               = "1.0.96"
+tokio                    = { version = "1.27.0", default-features = false, features = ["time"] }
+tracing                  = "0.1.37"
+unionlabs                = { workspace = true }
+
+[dev-dependencies]
+derive_more = { workspace = true, features = ["display"] }
+enumorph    = "0.1.2"
+
+[features]
+default = []
+
+arbitrary = ["dep:arbitrary", "unionlabs/arbitrary"]
