+++ voyager/voyager-message/fuzz_targets/serde_roundtrip.rs
@@ -1,8 +1,9 @@
 #![no_main]
 
 use libfuzzer_sys::fuzz_target;
+use queue_msg::QueueMsg;
 use unionlabs::test_utils::*;
-use voyager_message::{QueueMsg, RelayerMsgTypes};
+use voyager_message::RelayerMsgTypes;
 
 fuzz_target!(|data: QueueMsg<RelayerMsgTypes>| {
     assert_json_roundtrip(&data);
